Added User and Control classes
There are some explanations and questions in the code, there are lots of comments too. Feel free to ask or tell the team if you have any other ideas or if you disagree, or anything really. I have placed this in the main because it works and this way we have a working version already, future work should be done on develop